requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
lxml>=5.1.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data processing
requests>=2.31.0
pandas>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI

# Code quality
black>=23.12.0
flake8>=7.0.0
isort>=5.13.0
mypy>=1.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: Production monitoring
prometheus-client>=0.19.0